feature Refactor entries commands into a dedicated package, simplify `rootCmd` initialization, and modularize parsing and filtering logic.
